+++ README.md
@@ -1,1 +1,4 @@
-SQLFlow client in Java
+# SQLFlow client for Java Developers
+
+## Requirements
+ - Java 8+
