+++ pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>org.sqlflow</groupId>
-    <artifactId>sqlflowclient</artifactId>
+    <artifactId>jsqlflow</artifactId>
     <version>1.0-SNAPSHOT</version>
